fix(agents): address review feedback on PR #118588 - lint, types, tests, and split oversized model-selection
