+++ src/auto-reply/reply/get-reply.ts
@@ -73,7 +73,8 @@ import {
   hasInboundMediaForUnderstanding,
 } from "./inbound-media.js";
 import { emitPreAgentMessageHooks } from "./message-preprocess-hooks.js";
-import { createFastTestModelSelectionState, createModelSelectionState } from "./model-selection.js";
+import { createFastTestModelSelectionState } from "./model-selection-fast-test.js";
+import { createModelSelectionState } from "./model-selection.js";
 import {
   PENDING_FINAL_DELIVERY_CLEAR_PATCH,
   sanitizePendingFinalDeliveryText,
