+++ src/auto-reply/reply/model-selection.ts
@@ -13,7 +13,6 @@ import type { ModelCatalogEntry } from "../../agents/model-catalog.js";
 import type { ModelCatalogSnapshot } from "../../agents/model-catalog.types.js";
 import type { ModelFallbackRouteResolution } from "../../agents/model-fallback.types.js";
 import {
-  type ModelAliasIndex,
   buildConfiguredModelCatalog,
   legacyModelKey,
   modelKey,
@@ -48,6 +47,7 @@ export {
 } from "./model-selection-directive.js";
 export { resolveContextTokens } from "./model-selection-context.js";
 import { normalizeRuntimeRef, resolveRuntimeNormalization } from "./model-runtime-normalization.js";
+import type { ModelSelectionState, ThinkingDefaultSelection } from "./model-selection-fast-test.js";
 import {
   isStaleHeartbeatAutoFallbackOverride,
   normalizeStoredRuntimeModelRef,
@@ -57,67 +57,11 @@ import {
 
 type ModelCatalog = ModelCatalogEntry[];
 
-type ThinkingDefaultSelection = {
-  provider: string;
-  model: string;
-  agentRuntime?: string | null;
-};
-
-type ModelSelectionState = {
-  provider: string;
-  model: string;
-  requestedRouteResolution: ModelFallbackRouteResolution;
-  allowedModelKeys: Set<string>;
-  allowedModelCatalog: ModelCatalog;
-  policyAliasIndex: ModelAliasIndex;
-  resetModelOverride: boolean;
-  resetModelOverrideRef?: string;
-  resetModelOverrideReason?: "disallowed" | "stale" | "temporarily-unavailable";
-  modelPolicyConfigPath?: string;
-  modelPolicyRepairConfigPath?: string;
-  resolveThinkingCatalog: () => Promise<ModelCatalog | undefined>;
-  resolveDefaultThinkingLevel: (selection?: ThinkingDefaultSelection) => Promise<ThinkLevel>;
-  hasConfiguredThinkingDefault?: boolean;
-  /** Default reasoning level from model capability: "on" if model has reasoning, else "off". */
-  resolveDefaultReasoningLevel: () => Promise<"on" | "off">;
-  needsModelCatalog: boolean;
-  modelContextWindow?: number;
-  modelContextTokens?: number;
-};
-
 function resolveConfiguredModelThinkingDefault(raw: unknown): ThinkLevel | undefined {
   if (raw === false || raw === "disabled" || raw === "none") {
     return "off";
   }
   return typeof raw === "string" ? normalizeThinkLevel(raw) : undefined;
-}
-
-/** Creates minimal model-selection state for fast test mode. */
-export function createFastTestModelSelectionState(params: {
-  agentCfg: NonNullable<NonNullable<OpenClawConfig["agents"]>["defaults"]> | undefined;
-  provider: string;
-  model: string;
-}): ModelSelectionState {
-  return {
-    provider: params.provider,
-    model: params.model,
-    requestedRouteResolution: "resolved",
-    allowedModelKeys: new Set<string>(),
-    allowedModelCatalog: [],
-    policyAliasIndex: { byAlias: new Map(), byKey: new Map() },
-    resetModelOverride: false,
-    resetModelOverrideRef: undefined,
-    resetModelOverrideReason: undefined,
-    modelPolicyConfigPath: undefined,
-    modelPolicyRepairConfigPath: undefined,
-    resolveThinkingCatalog: async () => [],
-    resolveDefaultThinkingLevel: async () => params.agentCfg?.thinkingDefault as ThinkLevel,
-    hasConfiguredThinkingDefault: params.agentCfg?.thinkingDefault !== undefined,
-    resolveDefaultReasoningLevel: async () => "off",
-    needsModelCatalog: false,
-    modelContextWindow: undefined,
-    modelContextTokens: undefined,
-  };
 }
 
 const modelCatalogRuntimeLoader = createLazyImportLoader(
